refactor: rename new and touched files to kebab-case
Per CLAUDE.md file naming convention.
