+++ src/app.spec.tsx
@@ -2,7 +2,7 @@ import { act } from 'react';
 import { createRoot } from 'react-dom/client';
 import { describe, expect, it } from 'vitest';
 
-import App from './App';
+import App from './app';
 
 describe('App', () => {
   it('renders the game title', () => {
