+++ src/app.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Route, Routes } from 'react-router';
 
-import Home from '~/views/Home';
-import Game from '~/views/Game';
+import Home from '~/views/home';
+import Game from '~/views/game';
 
 export default function App() {
   return (
