+++ src/components/game-canvas.tsx
