+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
 import './index.css';
-import App from './App';
+import App from './app';
 
 const container = document.getElementById('root');
 if (!container) {
