+++ src/views/game.tsx
@@ -1,4 +1,4 @@
-import GameCanvas from '~/components/GameCanvas';
+import GameCanvas from '~/components/game-canvas';
 
 export default function Game() {
   return <GameCanvas className="h-screen w-screen" />;
